Better

+++ include/screen.h
@@ -2,7 +2,7 @@
 #include <buffers.h>
 #include <stdint.h>
 
-extern double ms_time_elapsed;
+extern uint64_t ms_time_elapsed;
 
 void screen_init(int w, int h);
 void screen_render();
